snap-confine: fallback gracefully on a cgroup v2 only system
On a cgroup v2 only system we cannot use certain features like
freezer or pid cgroups yet. In order to not die() when we run
on such systems this PR implements detection and skipping of
those.

This will have to be done "properly" but for now it ensures
that snaps continue to run on cgroup v2 only systems.